pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["user <bav-574@mail.ru>"]
readme = "README.md"
packages = [{include = "todo_list"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
aiohttp = "^3.8.4"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
